+++ composer.json
@@ -6,6 +6,7 @@
         "ext-ctype": "*",
         "ext-iconv": "*",
         "easycorp/easyadmin-bundle": "^2.3",
+        "nikic/php-parser": "^4.3",
         "symfony/asset": "4.3.*",
         "symfony/console": "4.3.*",
         "symfony/dotenv": "4.3.*",
@@ -18,6 +19,8 @@
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.3",
+        "phpstan/extension-installer": "^1.0",
+        "phpstan/phpstan-symfony": "^0.12.1",
         "squizlabs/php_codesniffer": "^3.5",
         "symfony/maker-bundle": "^1.13"
     },
