+++ src/Service/FilterService.php
@@ -13,9 +13,11 @@ class FilterService
     public function execute(Request $request): FiltersData
     {
         $filterData = new FiltersData();
+
         $filterData->setProgramName($request->get('type'));
         $filterData->setProgramPrice($request->get('price'));
         $filterData->setPage($request->get('page', 1));
+
         return $filterData;
     }
 }
