+++ symfony.lock
@@ -121,6 +121,15 @@
     "pagerfanta/pagerfanta": {
         "version": "v2.1.3"
     },
+    "phpstan/extension-installer": {
+        "version": "1.0.3"
+    },
+    "phpstan/phpstan": {
+        "version": "0.12.2"
+    },
+    "phpstan/phpstan-symfony": {
+        "version": "0.12.1"
+    },
     "psr/cache": {
         "version": "1.0.1"
     },
